Add hacky mouse-emulation-via-touch.
It's hacky because it has jumpy behavior in some edge cases, and touch
should be exposed properly via a touch API. Although perhaps its useful
to have a (good) mouse-emulation mode available for opt in.
One touch for MouseButton1.
Two touches for MouseButton2.
Three touches for both MouseButton1 and MouseButton2.