Implement new experimental MouseMovementCallback API.
On the browser side, it is implemented via movementX and movementY
extensions to MouseEvent interface. This works really well.
On the desktop site, for now, it is implemented by reusing
CursorPositionCallback and calculating delta pointer positions
ourselves. This can be later improved by implementing similar API in
GLFW (which would use raw mouse input on Windows and analogous apis on
other OSes to get the best mouse movement input possible).
This is an investigation of an experimental API addition related to
glfw/glfw#125; if it goes well I'll make a PR for glfw. /cc @elmindreda