Simplify Open to return io.ReadCloser.
This allows the browser implementation to return the underlying
io.ReaderCloser directly, without having to first cache the entire
thing. That enables streaming of the underlying source to be possible
(which is possible in browser after Fetch Transport, see
gopherjs/gopherjs#454.

If seeking is desired, it can be added by the user on top of the
returned io.ReadCloser. In many cases, seeking is not very critical and
easy to avoid, and definitely not worth the expensive overhead of
always supporting it.

This is a slightly breaking API change if you depend on seeking, but
easy to work around in client code.