implement focus callback in browser
Use the browser's focus and blur events to implement the focus callback,
which is informed when the browser window is in focus (active window)
or not.

Depending on whether or not we want to consider multiple tabs as "all
focused" or if each is focused individually, we may need to change the
listener to attach to the document instead of the window. The browser
functionality related to that has not been explored.

Fixes https://github.com/goxjs/glfw/issues/12.

GitHub-Pull-Request: https://github.com/goxjs/glfw/pull/14