implement close callback in browser
Implement the close callback via the browser's beforeunload event,
which is fired just prior to the page being unloaded.

It might be worth documenting there are some limitations on what the
callback handler can do with the DOM depending on the browser.
In some, opening a dialog is not allowed (to prevent annoying pop-ups).

GitHub-Pull-Request: https://github.com/goxjs/glfw/pull/15